# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "0.4.0",
    "description": "Contract for the current Next.js HTTP routes. Version follows the repository release line. This project is pre-1.0: request and response details may change between releases. See api-contract.md for supported guarantees and known gaps.",
    "title": "Nuera Quicksilver HTTP API",
    "summary": "Pre-1.0 API contract; not a stable compatibility promise."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/agents/drafts": {
      "post": {
        "operationId": "createAgentDraft",
        "tags": [
          "Agents"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/JsonResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/JsonObject"
        }
      }
    },
    "/api/agents/drafts/submit": {
      "post": {
        "operationId": "submitAgentDraft",
        "tags": [
          "Agents"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/JsonResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/JsonObject"
        }
      }
    },
    "/api/agents/publish": {
      "post": {
        "operationId": "publishAgent",
        "tags": [
          "Agents"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/JsonResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/JsonObject"
        }
      }
    },
    "/api/agents/review": {
      "post": {
        "operationId": "reviewAgent",
        "tags": [
          "Agents"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/JsonResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/JsonObject"
        }
      }
    },
    "/api/agents/rollback": {
      "post": {
        "operationId": "rollbackAgent",
        "tags": [
          "Agents"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/JsonResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/JsonObject"
        }
      }
    },
...
    "/api/workflows/drafts": {
      "post": {
        "operationId": "saveWorkflowDraft",
        "tags": [
          "Workflows"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/JsonResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/JsonObject"
        }
      }
    },
    "/api/workflows/drafts/submit": {
      "post": {
        "operationId": "submitWorkflowDraft",
        "tags": [
          "Workflows"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/JsonResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/JsonObject"
        }
      }
    },
...
    "/api/workflows/publish": {
      "post": {
        "operationId": "publishWorkflow",
        "tags": [
          "Workflows"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/JsonResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/JsonObject"
        }
      }
    },
    "/api/workflows/review": {
      "post": {
        "operationId": "reviewWorkflow",
        "tags": [
          "Workflows"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/JsonResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/JsonObject"
        }
      }
    },
    "/api/workflows/rollback": {
      "post": {
        "operationId": "rollbackWorkflow",
        "tags": [
          "Workflows"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/JsonResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/JsonObject"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "retryAfterSeconds": {
            "type": "integer",
            "minimum": 1
          },
          "code": {
            "type": "string"
          },
          "needs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "Common fields where emitted; not guaranteed across all routes.",
        "additionalProperties": true
      },
      "JsonObject": {
        "type": "object",
        "description": "Route-specific shape not exhaustively captured yet.",
        "additionalProperties": true
      },
...
    },
    "requestBodies": {
      "JsonObject": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/JsonObject"
            }
          }
        },
        "required": true
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "opaque principal or configured supervisor token",
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "RateLimitResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Rate limited by the current in-process token bucket.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        }
      },
      "JsonResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/JsonObject"
            }
          }
        },
        "description": "Successful JSON response; exact route-specific fields are not yet fully specified."
      },
      "ErrorResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Route-specific error; status and body vary by failure."
      },
...
    },
...
  }
}